# swagger/swaggerDocumentation.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.0.0",
    "title": "COM6504 Team 404",
    "description": "The assignment of team 404",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "http:/mission.lomofu.com",
      "description": "Testing server"
    }
  ],
  "paths": {
    "/api/story": {
      "post": {
        "tags": [
          "story"
        ],
        "description": "create a new story",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Story"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully create new story",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                },
                "example": {
                  "id": "623f4276936d27eccd4a2d42"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Miss the title value"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Server error, insert failed! Please try again!"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "story"
        ],
        "description": "get story details by id",
        "parameters": [{
          "name": "id",
          "in": "query",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "allowEmptyValue": true
        }],
        "responses": {
          "200": {
            "description": "Successfully get story details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                },
                "example": {
                  "id": "6262fe1cdbcd4061deb52a6b",
                  "title": "Find clues in Sheffield Library",
                  "author": "Luna",
                  "image": "/Desktop/library.jpg",
                  "description": "This is Diamond library at the University of Sheffield. I hid a clue under one of the desks in the library. Please find it."
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Miss the title value"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Server error, insert failed! Please try again!"
                }
              }
            }
          }
        }
      }
    },
    "/api/story/list": {
      "get": {
        "tags": [
          "story"
        ],
        "description": "get story list",
        "responses": {
          "200": {
            "description": "Successfully get stories list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryList"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Server error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "id": {
        "type": "string",
        "description": "story id",
        "example": "6262fe1cdbcd4061deb52a6b"
      },
      "title": {
        "type": "string",
        "description": "story title",
        "example": "Find clues in Sheffield Library"
      },
      "author": {
        "type": "string",
        "example": "Luna"
      },
      "image": {
        "type": "string",
        "description": "image url or uploaded",
        "example": "https://a0.muscache.com/im/pictures/09efb5fc-7bf1-44c0-b0ef-241a38533e67.jpg?im_w=1680"
      },
      "description": {
        "type": "string",
        "description": "story description",
        "example": "This is Diamond library at the University of Sheffield. I hid a clue under one of the desks in the library. Please find it."
      },
      "Story": {
        "type": "object",
        "properties": {
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "author": {
            "$ref": "#/components/schemas/author"
          },
          "image": {
            "$ref": "#/components/schemas/image"
          },
          "description": {
            "$ref": "#/components/schemas/description"
          }
        }
      },
      "StoryList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "story": {
              "$ref": "#/components/schemas/Story"
            }
          }
        }
      },
      "Error": {
        "type": "string",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Success": {
        "type": "string",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      }
    }
  }
}